+++ src/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: uphokaew <uphokaew@student.42bangkok.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/29 21:06:46 by uphokaew          #+#    #+#             */
-/*   Updated: 2024/01/26 14:49:05 by uphokaew         ###   ########.fr       */
+/*   Updated: 2024/01/26 15:06:49 by uphokaew         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ static int	specifier_check(const char specifier, va_list args)
 	else if (specifier == 'x' || specifier == 'X')
 		length += ft_printhex(va_arg(args, unsigned int), specifier);
 	else if (specifier == 'p')
-		length += ft_printptr(va_arg(args, unsigned long long));
+		length += ft_printptr(va_arg(args, uintptr_t));
 	else if (specifier == '%')
 		length += ft_printchr('%');
 	return (length);
